// client/typechain-types/Woofy.ts
export declare namespace Woofy {
  export type WoofyDataStruct = {
    tokenId: BigNumberish;
    tokenUri: string;
    price: BigNumberish;
    status: BigNumberish;
    owner: string;
  };

  export type WoofyDataStructOutput = [
    BigNumber,
    string,
    BigNumber,
    number,
    string
  ] & {
    tokenId: BigNumber;
    tokenUri: string;
    price: BigNumber;
    status: number;
    owner: string;
  };
}